wip: parallelize remote api call to avoid timeout
